Move icmd wrapper to internal/pkg/test/exec
Rework some of the code to make it more similar to exec.Command, so that
moving from exec.Command to this version is simpler.

Rework some of the code already using this.
